feat: Add support for movie series and parts
Enhances the movie data structure and UI to accommodate movies that are part of a series.
Introduces `seriesTitle` and `partNumber` fields to the `Movie` type.
Updates `MovieCard` to display "Part X" for multi-part series.
Adds logic to `MovieDetailsPage` to group and display other parts of a series.
Refactors bot commands to handle series data during movie addition.